gradio/requirements.txt
torch==1.7.1
torchvision==0.8.2
GitPython==3.1.43
gradio==4.44.1
huggingface-hub==0.28.1

Pillow==9.3.0
blobfile==3.1.0
mypy==1.14.1
numpy==1.24.3
pytest==8.3.5
requests==2.32.4
